move out to stability.go
