+++ tests/e2e/tidbcluster/tidbcluster.go
@@ -15,7 +15,6 @@ package tidbcluster
 
 import (
 	"context"
-	errors1 "errors"
 	"fmt"
 	_ "net/http/pprof"
 	"strconv"
@@ -34,7 +33,6 @@ import (
 	"github.com/pingcap/tidb-operator/pkg/manager/member"
 	"github.com/pingcap/tidb-operator/pkg/monitor/monitor"
 	"github.com/pingcap/tidb-operator/pkg/scheme"
-	"github.com/pingcap/tidb-operator/pkg/util"
 	tcconfig "github.com/pingcap/tidb-operator/pkg/util/config"
 	"github.com/pingcap/tidb-operator/tests"
 	e2econfig "github.com/pingcap/tidb-operator/tests/e2e/config"
@@ -46,7 +44,6 @@ import (
 	"github.com/pingcap/tidb-operator/tests/pkg/apimachinery"
 	"github.com/pingcap/tidb-operator/tests/pkg/blockwriter"
 	"github.com/pingcap/tidb-operator/tests/pkg/fixture"
-	"github.com/pingcap/tidb-operator/tests/pkg/mock"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
@@ -114,8 +111,8 @@ var _ = ginkgo.Describe("TiDBCluster", func() {
 		framework.ExpectNoError(err, "failed to create port forwarder")
 		fwCancel = cancel
 		cfg = e2econfig.TestConfig
-		OperatorFeatures := map[string]bool{"AutoScaling": true}
-		cfg.OperatorFeatures = OperatorFeatures
+		// OperatorFeatures := map[string]bool{"AutoScaling": true}
+		// cfg.OperatorFeatures = OperatorFeatures
 		ocfg = e2econfig.NewDefaultOperatorConfig(cfg)
 		if ocfg.Enabled(features.AdvancedStatefulSet) {
 			stsGetter = astsHelper.NewHijackClient(c, asCli).AppsV1()
@@ -929,495 +926,6 @@ var _ = ginkgo.Describe("TiDBCluster", func() {
 				return true, nil
 			})
 			framework.ExpectNoError(err, "not clear TiDB failureMembers when scale TiDB to zero")
-		})
-	})
-
-	// add [Stability] to prevent running auto scaling related cases for now
-	ginkgo.Context("[Stability][Feature: AutoScaling]", func() {
-		setCPUUsageAndQuota := func(tc *v1alpha1.TidbCluster, monitor *v1alpha1.TidbMonitor, usage, quota, memberType string, insts []string) {
-			// TODO: This duration is now hard-coded in PD
-			// It may become configurable in the future
-			duration := "60s"
-			mp := &mock.MonitorParams{
-				Name:                tc.Name,
-				KubernetesNamespace: tc.Namespace,
-				MemberType:          memberType,
-				Duration:            duration,
-				Value:               usage,
-				QueryType:           "cpu_usage",
-				InstancesPod:        insts,
-			}
-			err := mock.SetPrometheusResponse(monitor.Name, monitor.Namespace, mp, fw)
-			framework.ExpectNoError(err, "set %s cpu usage mock metrics error", memberType)
-
-			mp = &mock.MonitorParams{
-				Name:                tc.Name,
-				KubernetesNamespace: tc.Namespace,
-				MemberType:          memberType,
-				Duration:            duration,
-				Value:               quota,
-				QueryType:           "cpu_quota",
-				InstancesPod:        insts,
-			}
-			err = mock.SetPrometheusResponse(monitor.Name, monitor.Namespace, mp, fw)
-			framework.ExpectNoError(err, "set %s cpu quota mock metrics error", memberType)
-		}
-		ginkgo.It("should auto scale TiKV pods", func() {
-			ginkgo.By("Deploy initial tc")
-			clusterName := "auto-scaling-tikv"
-			tc := fixture.GetTidbCluster(ns, clusterName, utilimage.TiDBNightlyVersion)
-			tc.Spec.PD.Replicas = 1
-			tc.Spec.TiKV.Replicas = 3
-			tc.Spec.PD.Config.Set("pd-server.metric-storage", "http://monitor-prometheus:9090")
-
-			_, err := cli.PingcapV1alpha1().TidbClusters(ns).Create(tc)
-			framework.ExpectNoError(err, "Create TidbCluster error")
-			err = oa.WaitForTidbClusterReady(tc, 30*time.Minute, 15*time.Second)
-			framework.ExpectNoError(err, "Check TidbCluster error")
-
-			ginkgo.By("Create tidb monitor with e2e image")
-			a := e2econfig.TestConfig.E2EImage
-			colonIdx := strings.LastIndexByte(a, ':')
-			image := a[:colonIdx]
-			tag := a[colonIdx+1:]
-			monitor := fixture.NewTidbMonitor("monitor", ns, tc, false, false, false)
-			monitor.Spec.Prometheus.BaseImage = image
-			monitor.Spec.Prometheus.Version = tag
-
-			_, err = cli.PingcapV1alpha1().TidbMonitors(ns).Create(monitor)
-			framework.ExpectNoError(err, "Create TidbMonitor error")
-			err = tests.CheckTidbMonitor(monitor, cli, c, fw)
-			framework.ExpectNoError(err, "Check TidbMonitor error")
-
-			ginkgo.By("Create TidbClusterAutoScaler")
-			tcas := fixture.GetTidbClusterAutoScaler("auto-scaler", ns, tc, monitor)
-			tcas.Spec.TiKV = &v1alpha1.TikvAutoScalerSpec{}
-			tcas.Spec.TiKV.ScaleInIntervalSeconds = pointer.Int32Ptr(1)
-			tcas.Spec.TiKV.ScaleOutIntervalSeconds = pointer.Int32Ptr(1)
-			tcas.Spec.TiKV.Resources = map[string]v1alpha1.AutoResource{
-				"storage": {
-					CPU:     resource.MustParse("1024m"),
-					Memory:  resource.MustParse("2Gi"),
-					Storage: resource.MustParse("10Gi"),
-					Count:   pointer.Int32Ptr(2),
-				},
-			}
-			tcas.Spec.TiKV.Rules = map[v1.ResourceName]v1alpha1.AutoRule{
-				v1.ResourceCPU: {
-					MaxThreshold:  0.5,
-					MinThreshold:  pointer.Float64Ptr(0.2),
-					ResourceTypes: []string{"storage"},
-				},
-			}
-			_, err = cli.PingcapV1alpha1().TidbClusterAutoScalers(ns).Create(tcas)
-			framework.ExpectNoError(err, "Create TidbClusterAutoScaler error")
-
-			pdClient, cancel, err := proxiedpdclient.NewProxiedPDClient(c, fw, ns, clusterName, false)
-			framework.ExpectNoError(err, "create pdapi error")
-			defer cancel()
-
-			var autoTc v1alpha1.TidbCluster
-			autoTcListOption := metav1.ListOptions{
-				LabelSelector: labels.SelectorFromSet(labels.Set{
-					label.AutoInstanceLabelKey: tcas.Name,
-					label.BaseTCLabelKey:       tc.Name,
-				}).String(),
-			}
-
-			// TiKV autoscaling
-			baseTiKVs := make([]string, 0, tc.Spec.TiKV.Replicas)
-			for i := int32(0); i < tc.Spec.TiKV.Replicas; i++ {
-				baseTiKVs = append(baseTiKVs, util.GetPodName(tc, v1alpha1.TiKVMemberType, i))
-			}
-			var autoTiKV string
-			ginkgo.By("Case 1: No autoscaling cluster and CPU usage over max threshold")
-			setCPUUsageAndQuota(tc, monitor, "35.0", "1.0", v1alpha1.TiKVMemberType.String(), baseTiKVs)
-			// A new cluster should be created and all TiKV stores are up
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcList, err := cli.PingcapV1alpha1().TidbClusters(tc.Namespace).List(autoTcListOption)
-				if err != nil {
-					log.Logf("failed to list tc %q with selector %+v", tc.Name, autoTcListOption)
-					return false, err
-				}
-
-				if len(tcList.Items) < 1 {
-					log.Logf("autoscaling tikv cluster is not created")
-					return false, nil
-				}
-
-				autoTc = tcList.Items[0]
-
-				if autoTc.Spec.TiKV == nil {
-					log.Logf("the created cluster has no tikv spec")
-					return false, errors1.New("the created cluster has no tikv spec")
-				}
-
-				autoTiKV = util.GetPodName(&autoTc, v1alpha1.TiKVMemberType, 0)
-				setCPUUsageAndQuota(tc, monitor, "20.0", "1.0", v1alpha1.TiKVMemberType.String(), append(baseTiKVs, autoTiKV))
-
-				if len(autoTc.Status.TiKV.Stores) < int(autoTc.Spec.TiKV.Replicas) {
-					log.Logf("len(.Status.TiKV.Stores) < .Spec.TiKV.Replicas for tc %q", autoTc.Name)
-					return false, nil
-				}
-
-				for _, store := range autoTc.Status.TiKV.Stores {
-					if store.State != v1alpha1.TiKVStateUp {
-						log.Logf("autoscaling tikv cluster not ready, store %s is not %s", store.PodName, v1alpha1.TiKVStateUp)
-						return false, nil
-					}
-				}
-
-				storeID := ""
-				podNameTikv0 := util.GetPodName(&autoTc, v1alpha1.TiKVMemberType, int32(0))
-				for k, v := range autoTc.Status.TiKV.Stores {
-					if v.PodName == podNameTikv0 {
-						storeID = k
-						break
-					}
-				}
-				if storeID == "" {
-					log.Logf("no store found for pod %q", podNameTikv0)
-					return false, nil
-				}
-				sid, err := strconv.ParseUint(storeID, 10, 64)
-				if err != nil {
-					log.Logf("failed to parse store id %q for pod %q", storeID, podNameTikv0)
-					return false, err
-				}
-				info, err := pdClient.GetStore(sid)
-				if err != nil {
-					log.Logf("fail to get stores")
-					return false, err
-				}
-
-				// Check labels
-				expectedLabels := map[string]string{
-					"specialUse":    "hotRegion",
-					"resource-type": "storage",
-					"group":         "pd-auto-scaling-tikv", // This label is subject to change
-				}
-				for _, label := range info.Store.Labels {
-					if value, ok := expectedLabels[label.Key]; ok && value != label.Value {
-						errMsg := fmt.Sprintf("expected label %s of tc[%s/%s]'s store %d to have value %s, got %s", label.Key, autoTc.Namespace, autoTc.Name, sid, expectedLabels[label.Key], label.Value)
-						log.Logf(errMsg)
-						return false, fmt.Errorf(errMsg)
-					}
-				}
-
-				return true, nil
-			})
-			framework.ExpectNoError(err, "check create autoscaling tikv cluster error")
-			log.Logf("success to check create autoscaling tikv cluster")
-
-			ginkgo.By("Case 2: Has an autoscaling cluster and CPU usage between max threshold and min threshold")
-			setCPUUsageAndQuota(tc, monitor, "20.0", "1.0", v1alpha1.TiKVMemberType.String(), append(baseTiKVs, autoTiKV))
-			// The TiKV replicas should remain unchanged
-			err = wait.Poll(10*time.Second, 3*time.Minute, func() (done bool, err error) {
-				tcPtr, err := cli.PingcapV1alpha1().TidbClusters(autoTc.Namespace).Get(autoTc.Name, metav1.GetOptions{})
-				if err != nil {
-					log.Logf("failed to get tc %q", tcPtr.Name)
-					return false, err
-				}
-
-				autoTc = *tcPtr
-
-				if autoTc.Spec.TiKV.Replicas != 1 {
-					log.Logf("expected tc[%s/%s]'s tikv replicas to stay at 1, now %d", autoTc.Namespace, autoTc.Name, autoTc.Spec.TiKV.Replicas)
-					return true, nil
-				}
-
-				log.Logf("tikv is not autoscaled")
-				return false, nil
-			})
-			framework.ExpectEqual(err, wait.ErrWaitTimeout, "expect tikv is not scaled under normal utilization for 3 minutes")
-
-			ginkgo.By("Case 3: Has an autoscaling cluster and CPU usage over max threshold")
-			setCPUUsageAndQuota(tc, monitor, "35.0", "1.0", v1alpha1.TiKVMemberType.String(), append(baseTiKVs, autoTiKV))
-			// The existing autoscaling cluster should be scaled out
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcPtr, err := cli.PingcapV1alpha1().TidbClusters(autoTc.Namespace).Get(autoTc.Name, metav1.GetOptions{})
-				if err != nil {
-					log.Logf("failed to get tc %q", tcPtr.Name)
-					return false, err
-				}
-
-				autoTc = *tcPtr
-
-				if autoTc.Spec.TiKV.Replicas < 2 {
-					log.Logf("autoscaling tikv cluster is not scaled out")
-					return false, nil
-				}
-
-				if len(autoTc.Status.TiKV.Stores) < int(autoTc.Spec.TiKV.Replicas) {
-					log.Logf("len(.Status.TiKV.Stores) < .Spec.TiKV.Replicas for tc %q", autoTc.Name)
-					return false, nil
-				}
-
-				for _, store := range autoTc.Status.TiKV.Stores {
-					if store.State != v1alpha1.TiKVStateUp {
-						log.Logf("autoscaling tikv cluster scaled out but store %s is not %s", store.PodName, v1alpha1.TiKVStateUp)
-						return false, nil
-					}
-				}
-
-				return true, nil
-			})
-			framework.ExpectNoError(err, "check scale out existing autoscaling tikv cluster error")
-			log.Logf("success to check scale out existing autoscaling tikv cluster")
-
-			pods := make([]string, len(baseTiKVs))
-			copy(pods, baseTiKVs)
-			for i := int32(0); i < autoTc.Spec.TiKV.Replicas; i++ {
-				pods = append(pods, util.GetPodName(&autoTc, v1alpha1.TiKVMemberType, i))
-			}
-
-			ginkgo.By("Case 4: CPU usage below min threshold")
-			setCPUUsageAndQuota(tc, monitor, "0.0", "1.0", v1alpha1.TiKVMemberType.String(), pods)
-			// The autoscaling cluster should be scaled in
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcPtr, err := cli.PingcapV1alpha1().TidbClusters(autoTc.Namespace).Get(autoTc.Name, metav1.GetOptions{})
-				if err != nil {
-					if errors.IsNotFound(err) {
-						log.Logf("tc %q not found", tcPtr.Name)
-						return true, nil
-					}
-					log.Logf("failed to get tc %q", tcPtr.Name)
-					return false, err
-				}
-
-				autoTc = *tcPtr
-
-				if autoTc.Spec.TiKV.Replicas > 1 {
-					log.Logf("autoscaling tikv cluster is not scaled in, replicas=%d", autoTc.Spec.TiKV.Replicas)
-					return false, nil
-				}
-
-				if autoTc.Spec.TiKV.Replicas <= 1 {
-					log.Logf("autoscaling tikv cluster tc[%s/%s] is scaled in", autoTc.Namespace, autoTc.Name)
-					return true, nil
-				}
-
-				return false, nil
-			})
-
-			framework.ExpectNoError(err, "failed to check scale in autoscaling tikv cluster")
-			log.Logf("success to check scale in autoscaling tikv cluster")
-
-			ginkgo.By("Case 5: CPU usage below min threshold for a long time")
-			// The autoscaling cluster should be deleted
-			err = wait.Poll(5*time.Second, 10*time.Minute, func() (done bool, err error) {
-				tcList, err := cli.PingcapV1alpha1().TidbClusters(tc.Namespace).List(autoTcListOption)
-				if err != nil {
-					log.Logf("failed to list tc %q with selector %+v", tc.Name, autoTcListOption)
-					return false, err
-				}
-
-				if len(tcList.Items) > 0 {
-					log.Logf("autoscaling tikv cluster is not deleted")
-					return false, nil
-				}
-
-				log.Logf("autoscaling tikv cluster deleted")
-				return true, nil
-			})
-			framework.ExpectNoError(err, "check delete autoscaling tikv cluster error")
-			log.Logf("success to check delete autoscaling tikv cluster")
-		})
-
-		ginkgo.It("should auto scale TiDB pods", func() {
-			ginkgo.By("Deploy initial tc")
-			clusterName := "auto-scaling-tidb"
-			tc := fixture.GetTidbCluster(ns, clusterName, utilimage.TiDBNightlyVersion)
-			tc.Spec.PD.Replicas = 1
-			tc.Spec.TiDB.Replicas = 2
-			tc.Spec.TiKV.Replicas = 3
-			tc.Spec.PD.Config.Set("pd-server.metric-storage", "http://monitor-prometheus:9090")
-
-			_, err := cli.PingcapV1alpha1().TidbClusters(ns).Create(tc)
-			framework.ExpectNoError(err, "Create TidbCluster error")
-			err = oa.WaitForTidbClusterReady(tc, 30*time.Minute, 15*time.Second)
-			framework.ExpectNoError(err, "Check TidbCluster error")
-
-			ginkgo.By("Create tidb monitor with e2e image")
-			a := e2econfig.TestConfig.E2EImage
-			colonIdx := strings.LastIndexByte(a, ':')
-			image := a[:colonIdx]
-			tag := a[colonIdx+1:]
-			monitor := fixture.NewTidbMonitor("monitor", ns, tc, false, false, false)
-			monitor.Spec.Prometheus.BaseImage = image
-			monitor.Spec.Prometheus.Version = tag
-
-			_, err = cli.PingcapV1alpha1().TidbMonitors(ns).Create(monitor)
-			framework.ExpectNoError(err, "Create TidbMonitor error")
-			err = tests.CheckTidbMonitor(monitor, cli, c, fw)
-			framework.ExpectNoError(err, "Check TidbMonitor error")
-			tcas := fixture.GetTidbClusterAutoScaler("auto-scaler", ns, tc, monitor)
-
-			ginkgo.By("Create TidbClusterAutoScaler")
-			tcas.Spec.TiDB = &v1alpha1.TidbAutoScalerSpec{}
-			tcas.Spec.TiDB.ScaleInIntervalSeconds = pointer.Int32Ptr(1)
-			tcas.Spec.TiDB.ScaleOutIntervalSeconds = pointer.Int32Ptr(1)
-			tcas.Spec.TiDB.Resources = map[string]v1alpha1.AutoResource{
-				"compute": {
-					CPU:    resource.MustParse("1024m"),
-					Memory: resource.MustParse("2Gi"),
-					Count:  pointer.Int32Ptr(2),
-				},
-			}
-			tcas.Spec.TiDB.Rules = map[v1.ResourceName]v1alpha1.AutoRule{
-				v1.ResourceCPU: {
-					MaxThreshold:  0.5,
-					MinThreshold:  pointer.Float64Ptr(0.2),
-					ResourceTypes: []string{"compute"},
-				},
-			}
-			_, err = cli.PingcapV1alpha1().TidbClusterAutoScalers(ns).Create(tcas)
-			framework.ExpectNoError(err, "Create TidbClusterAutoScaler error")
-
-			var autoTc v1alpha1.TidbCluster
-			autoTcListOption := metav1.ListOptions{
-				LabelSelector: labels.SelectorFromSet(labels.Set{
-					label.AutoInstanceLabelKey: tcas.Name,
-					label.BaseTCLabelKey:       tc.Name,
-				}).String(),
-			}
-
-			// TiDB Autoscaling
-			baseTiDBs := make([]string, 0, tc.Spec.TiDB.Replicas)
-			for i := int32(0); i < tc.Spec.TiDB.Replicas; i++ {
-				baseTiDBs = append(baseTiDBs, util.GetPodName(tc, v1alpha1.TiDBMemberType, i))
-			}
-			var autoTiDB string
-
-			ginkgo.By("Case 1: No autoscaling cluster and CPU usage over max threshold")
-			setCPUUsageAndQuota(tc, monitor, "35.0", "1.0", v1alpha1.TiDBMemberType.String(), baseTiDBs)
-			// A new cluster should be created
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcList, err := cli.PingcapV1alpha1().TidbClusters(tc.Namespace).List(autoTcListOption)
-				if err != nil {
-					log.Logf("failed to list tc with selector %+v", autoTcListOption)
-					return false, err
-				}
-
-				if len(tcList.Items) < 1 {
-					log.Logf("autoscaling tidb cluster is not created")
-					return false, nil
-				}
-
-				autoTc = tcList.Items[0]
-				autoTiDB = util.GetPodName(&autoTc, v1alpha1.TiDBMemberType, 0)
-				setCPUUsageAndQuota(tc, monitor, "20.0", "1.0", v1alpha1.TiDBMemberType.String(), append(baseTiDBs, autoTiDB))
-				return true, nil
-			})
-			framework.ExpectNoError(err, "check create autoscaling tidb cluster error")
-			log.Logf("success to check create autoscaling tidb cluster")
-
-			autoTiDB = util.GetPodName(&autoTc, v1alpha1.TiDBMemberType, 0)
-			ginkgo.By("Case 2: Has an autoscaling cluster and CPU usage between max threshold and min threshold")
-			setCPUUsageAndQuota(tc, monitor, "20.0", "1.0", v1alpha1.TiDBMemberType.String(), append(baseTiDBs, autoTiDB))
-			// The TiDB replicas should remain unchanged
-			err = wait.Poll(5*time.Second, 3*time.Minute, func() (done bool, err error) {
-				tcPtr, err := cli.PingcapV1alpha1().TidbClusters(autoTc.Namespace).Get(autoTc.Name, metav1.GetOptions{})
-				if err != nil {
-					return false, err
-				}
-
-				autoTc = *tcPtr
-
-				if autoTc.Spec.TiDB.Replicas != 1 {
-					log.Logf("expected tc[%s/%s]'s tidb replicas to stay at 1, now %d", autoTc.Namespace, autoTc.Name, autoTc.Spec.TiDB.Replicas)
-					return true, nil
-				}
-
-				log.Logf("confirm autoscaling tidb is not scaled when normal utilization")
-				return false, nil
-			})
-			framework.ExpectEqual(err, wait.ErrWaitTimeout, "expect tidb is not scaled when normal utilization for 5 minutes")
-
-			ginkgo.By("Case 3: Has an autoscaling cluster and CPU usage over max threshold")
-			setCPUUsageAndQuota(tc, monitor, "35.0", "1.0", v1alpha1.TiDBMemberType.String(), append(baseTiDBs, autoTiDB))
-			// The existing autoscaling cluster should be scaled out
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcPtr, err := cli.PingcapV1alpha1().TidbClusters(autoTc.Namespace).Get(autoTc.Name, metav1.GetOptions{})
-				if err != nil {
-					log.Logf("failed to get tc %q", tcPtr.Name)
-					return false, err
-				}
-
-				autoTc = *tcPtr
-
-				if autoTc.Spec.TiDB.Replicas < 2 {
-					log.Logf("autoscaling tidb cluster is not scaled out")
-					return false, nil
-				}
-
-				return true, nil
-			})
-			framework.ExpectNoError(err, "check scale out existing autoscaling tidb cluster error")
-			log.Logf("success to check scale out existing autoscaling tidb cluster")
-
-			pods := make([]string, len(baseTiDBs))
-			copy(pods, baseTiDBs)
-			for i := int32(0); i < autoTc.Spec.TiDB.Replicas; i++ {
-				pods = append(pods, util.GetPodName(&autoTc, v1alpha1.TiDBMemberType, i))
-			}
-
-			ginkgo.By("Case 4: CPU usage below min threshold")
-			setCPUUsageAndQuota(tc, monitor, "0.0", "1.0", v1alpha1.TiDBMemberType.String(), pods)
-			// The autoscaling cluster should be scaled in
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcPtr, err := cli.PingcapV1alpha1().TidbClusters(autoTc.Namespace).Get(autoTc.Name, metav1.GetOptions{})
-
-				if err != nil {
-					if errors.IsNotFound(err) {
-						log.Logf("tc %q not found", tcPtr.Name)
-						return true, nil
-					}
-					log.Logf("failed to get tc %q", tcPtr.Name)
-					return false, err
-				}
-
-				autoTc = *tcPtr
-
-				if autoTc.Spec.TiDB.Replicas > 1 {
-					log.Logf("autoscaling tidb cluster is not scaled in, replicas=%d", autoTc.Spec.TiDB.Replicas)
-					return false, nil
-				}
-
-				if autoTc.Spec.TiDB.Replicas <= 1 {
-					log.Logf("autoscaling tidb cluster tc[%s/%s] is scaled in", autoTc.Namespace, autoTc.Name)
-					return true, nil
-				}
-
-				return false, nil
-			})
-
-			framework.ExpectNoError(err, "failed to check scale in autoscaling tidb cluster")
-			log.Logf("success to check scale in autoscaling tidb cluster")
-
-			ginkgo.By("Case 5: CPU usage below min threshold for a long time")
-			// The autoscaling cluster should be deleted
-			err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
-				tcList, err := cli.PingcapV1alpha1().TidbClusters(tc.Namespace).List(autoTcListOption)
-				if err != nil {
-					log.Logf("failed to list tc with selector %+v", autoTcListOption)
-					return false, err
-				}
-
-				if len(tcList.Items) > 0 {
-					log.Logf("autoscaling tidb cluster is not deleted")
-					return false, nil
-				}
-
-				log.Logf("autoscaling tidb cluster deleted")
-				return true, nil
-			})
-			framework.ExpectNoError(err, "check delete autoscaling tidb cluster error")
-			log.Logf("success to check delete autoscaling tidb cluster")
-
-			// Clean autoscaler
-			err = cli.PingcapV1alpha1().TidbClusterAutoScalers(tcas.Namespace).Delete(tcas.Name, &metav1.DeleteOptions{})
-			framework.ExpectNoError(err, "failed to delete auto-scaler")
 		})
 	})
